update pattern in logback.xml
update

6 jvm

fix

fix

re-organize method tests

address feedback
